requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.10
pgvector==0.3.2
boto3==1.35.0
tos==2.9.0
Pillow==10.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.5.2
python-multipart==0.0.12
pytest==8.3.3
httpx==0.27.2
openai==1.54.4
dashscope==1.24.0
pytest-mock==3.14.0
celery==5.4.0
redis==5.1.1
